OnlineLogBinning

Documentation for OnlineLogBinning.

OnlineLogBinning.BinningAccumulator

OnlineLogBinning.LevelAccumulator

OnlineLogBinning.PairAccumulator

Base.length

Base.push!

Base.push!

Base.push!

OnlineLogBinning.Svalue

OnlineLogBinning.Svalue

OnlineLogBinning.Tvalue

OnlineLogBinning.Tvalue

OnlineLogBinning.bin_depth

OnlineLogBinning.reset!

OnlineLogBinning.update_Svalue!

OnlineLogBinning.update_Tvalue!

OnlineLogBinning.update_num_bins!

Statistics.mean

Statistics.var

OnlineLogBinning.BinningAccumulator — Type

BinningAccumulator{T}() where {T <: Number}

Main data structure for the binning analysis.

Contents

LvlAccums::Vector{LevelAccumulator{T}}

A wrapper around the LevelAccumulators from each binning level

OnlineLogBinning.LevelAccumulator — Type

LevelAccumulator{T <: Number}

Accumulator structure for a given binning level.

Contents

level::Int

Registers the binning level this accumulator is assigned

num_bins::Int

How many elements (i.e. bins) have been added to this accumulator

Taccum::T

Stands for Total Accumulator.

This represents the T accumulator for the mean: mean ≡ T / num_bins.

Saccum::T

Stands for Square Accumulator.

This represents the S accumulator for the variance: var ≡ S/(num_bins - 1).

Paccum::PairAccumulator{T}

An outward facing PairAccumulator to meet incoming data streams.

This accumulator processes the incoming data and then exports the Tvalue and Svalue into updates for Taccum and Saccum, respectively.

OnlineLogBinning.PairAccumulator — Type

PairAccumulator{T <: Number}

Accumulator that directly faces an incoming data stream. Two values from that stream enter and are processed into the exported values of Tvalue and Svalue.

Contents

fullpair::Bool

A Boolean to keep track of which element of the pair is being accessed. Additionally, when fullpair == true then the contents are exported.

values::MVector{2, T}

The individual values taken from the data stream to be processed. Both Tvalue and Svalue rely on them being accessible.

Base.length — Method

length(bacc::BinningAccumulator)

Return the number of LevelAccumulators there are.

Base.push! — Method

push!(bacc::BinningAccumulator, itr)

push! each value of the data stream itr through the BinningAccumulator.

Base.push! — Method

push!(bacc::BinningAccumulator, value::Number)

Add a single value from the data stream into the online binning analysis. The single value enters at the bin at the lowest level.

Base.push! — Method

push!(pacc::PairAccumulator, value::Number)

Overload Base.push! for a PairAccumulator. One can only push! a single value <: Number at a time into this type of accumulator.

OnlineLogBinning.Svalue — Method

Svalue(lacc::LevelAccumulator)

Function to calculate the online $S_{1,m+2}$ summation as:

\[S_{1,m+2} = S_{1,m} + S_{m+1,m+2} + \frac{m}{2(m+2)}\left( \frac{2}{m} T_{1,m} - T_{m+1,m+2} \right)^2.\]

where $T_{m+1,m+2}$ is the pairwise Tvalue for the PairAccumulator.

OnlineLogBinning.Svalue — Method

Svalue(pacc::PairAccumulator)

The $S$ function for a single pair following the accumulation of $m$ data points follows as

\[\begin{aligned} S_{m+1, m+2} &\equiv \sum_{k = m+1}^{m+2} \left( x_k - \frac{1}{2} T_{m+1,m+2} \right)^2 \\ &= \frac{1}{2}\left( x_{m+2} - x_{m+1} \right)^2. \end{aligned}\]

Thus, $S_{m+1,m+2}$ does not need to take $T_{m+1,m+2}$ as an argument.

OnlineLogBinning.Tvalue — Method

Tvalue(lacc::LevelAccumulator)

Function to calculate the online $T_{1,m+2}$ summation as:

\[T_{1,m+2} = T_{1,m} + T_{m+1,m+2},\]

where $T_{m+1,m+2}$ is the pairwise Tvalue for the PairAccumulator.

OnlineLogBinning.Tvalue — Method

Tvalue(pacc::PairAccumulator)

The $T$ function for a single pair following the accumulation of $m$ data points follows as

\[T_{m+1, m+2} \equiv \sum_{k = m+1}^{m+2} x_k = x_{m+1} + x_{m+2},\]

as expected.

OnlineLogBinning.bin_depth — Method

bin_depth(bacc::BinningAccumulator)

Number of binned levels present. length of the [BinningAccumulator] minus 1.

OnlineLogBinning.reset! — Method

reset!(pacc::PairAccumulator)

Return the PairAccumulator to its initial state. Presumably one just exported the Tvalue and Svalue from it before the reset!.

OnlineLogBinning.update_Svalue! — Method

update_Svalue!(lacc::LevelAccumulator)

Apply the Svalue formula to update lacc.Saccum and return the PairAccumulator increment.

OnlineLogBinning.update_Tvalue! — Method

update_Tvalue!(lacc::LevelAccumulator)

Apply the Tvalue formula to update lacc.Taccum and return the PairAccumulator increment.

OnlineLogBinning.update_num_bins! — Function

update_num_bins!(lacc::LevelAccumulator, [incr = 2])

Increment the number of bins accumulated by incr.

Statistics.mean — Method

mean( acc::LevelAccumulator )

Online measurement of the data stream mean.

Statistics.var — Method

var( acc::LevelAccumulator )

Online measurement of the data stream variance.