OnlineLogBinning

Documentation for OnlineLogBinning.

OnlineLogBinning.LevelAccumulator

OnlineLogBinning.PairAccumulator

Base.push!

OnlineLogBinning.Svalue

OnlineLogBinning.Tvalue

Statistics.mean

Statistics.var

OnlineLogBinning.LevelAccumulator — Type

LevelAccumulator{T <: Number}

Accumulator structure for a given binning level.

Contents

count::Bool

Keeps track of how many elements have been added to Baccum

level::Int

Registers the binning level this accumulator is assigned

nelements::Int

How many elements have been added to this accumulator

Baccum::T

Stands for Bare Accumulator.

New data pushed to this accumulator first passes into here.

Once count == true, then this accumulator is then passed onto Baccum in the next binning level. Additionally, it is also sent on to Taccum and Baccum.

Taccum::T

Stands for Total Accumulator.

This represents the T accumulator for the mean: mean ≡ T / nelements.

Saccum::T

Stands for Square Accumulator.

This represents the S accumulator for the variance: var ≡ S/(nelements - 1).

OnlineLogBinning.PairAccumulator — Type

PairAccumulator{T <: Number}

Base.push! — Method

push!(pacc::PairAccumulator, value::Number)

Overload Base.push! for a PairAccumulator. One can only push! a single value <: Number at a time into this type of accumulator.

OnlineLogBinning.Svalue — Method

Svalue(pacc::PairAccumulator)

The $S$ function for a single pair following the accumulation of $m$ data points follows as

\[S_{m+1, m+2} \equiv \sum_{k = m+1}^{m+2} \left( x_k - \frac{1}{2} T_{m+1,m+2} \right)^2.\]

Clearly, $S_{m+1,m+2}$ must be called following $T_{m+1,m+2}$.

OnlineLogBinning.Tvalue — Method

Tvalue(pacc::PairAccumulator)

The $T$ function for a single pair following the accumulation of $m$ data points follows as

\[T_{m+1, m+2} \equiv \sum_{k = m+1}^{m+2} x_k = x_{m+1} + x_{m+2},\]

as expected.

Statistics.mean — Method

mean( acc::LevelAccumulator )

Online measurement of the data stream mean.

Statistics.var — Method

var( acc::LevelAccumulator )

Online measurement of the data stream variance.