OnlineLogBinning

Documentation for OnlineLogBinning.

OnlineLogBinning.OLB_tested_numbers

OnlineLogBinning.BinningAccumulator

OnlineLogBinning.LevelAccumulator

OnlineLogBinning.PairAccumulator

Base.eltype

Base.length

Base.push!

Base.push!

Base.push!

Base.show

Base.show

OnlineLogBinning.Svalue

OnlineLogBinning.Svalue

OnlineLogBinning.Tvalue

OnlineLogBinning.Tvalue

OnlineLogBinning.bin_depth

OnlineLogBinning.binning_level

OnlineLogBinning.reset!

OnlineLogBinning.reset!

OnlineLogBinning.std_error

OnlineLogBinning.std_error

OnlineLogBinning.update_SandT!

OnlineLogBinning.update_num_bins!

OnlineLogBinning.var_of_mean

OnlineLogBinning.var_of_mean

Statistics.mean

Statistics.mean

Statistics.std

Statistics.std

Statistics.var

Statistics.var

OnlineLogBinning.OLB_tested_numbers — Constant

OLB_tested_numbers

Defines the list of tested numerical types for OnlineLogBinning.jl.

OnlineLogBinning.BinningAccumulator — Type

BinningAccumulator{T}() where {T <: Number}

Main data structure for the binning analysis. T == Float64 by default in the empty constructor.

Contents

LvlAccums::Vector{LevelAccumulator{T}}

A wrapper around the LevelAccumulators from each binning level

OnlineLogBinning.LevelAccumulator — Type

LevelAccumulator{T <: Number}

Accumulator structure for a given binning level.

Contents

level::Int

Registers the binning level this accumulator is assigned

num_bins::Int

How many elements (i.e. bins) have been added to this accumulator

Taccum::T

Stands for Total Accumulator.

This represents the T accumulator for the mean: mean ≡ T / num_bins.

Saccum::T

Stands for Square Accumulator.

This represents the S accumulator for the variance: var ≡ S/(num_bins - 1).

Paccum::PairAccumulator{T}

An outward facing PairAccumulator to meet incoming data streams.

This accumulator processes the incoming data and then exports the Tvalue and Svalue into updates for Taccum and Saccum, respectively.

OnlineLogBinning.PairAccumulator — Type

PairAccumulator{T <: Number}

Accumulator that directly faces an incoming data stream. Two values from that stream enter and are processed into the exported values of Tvalue and Svalue.

Contents

fullpair::Bool

A Boolean to keep track of which element of the pair is being accessed. Additionally, when fullpair == true then the contents are exported.

values::MVector{2, T}

The individual values taken from the data stream to be processed. Both Tvalue and Svalue rely on them being accessible.

Base.eltype — Method

eltype(::BinningAccumulator{T}) → T

Returns the type parameter for the BinningAccumulator.

Base.length — Method

length(bacc::BinningAccumulator)

Return the number of LevelAccumulators there are.

Base.push! — Method

push!(bacc::BinningAccumulator, itr)

push! each value of the data stream itr through the BinningAccumulator.

Base.push! — Method

push!(bacc::BinningAccumulator, value::Number)

Add a single value from the data stream into the online binning analysis. The single value enters at the bin at the lowest level.

Base.push! — Method

push!(pacc::PairAccumulator, value::Number)

Overload Base.push! for a PairAccumulator. One can only push! a single value <: Number at a time into this type of accumulator.

Base.show — Method

show([io::IO = stdout], bacc::BinningAccumulator)

Overload the Base.show function for human-readable displays.

Base.show — Method

show([io = stdout], lacc::LevelAccumulator)

Overload Base.show for human-readable displays.

OnlineLogBinning.Svalue — Method

Svalue(lacc::LevelAccumulator)

Function to calculate the online $S_{1,m+2}$ summation as:

\[S_{1,m+2} = S_{1,m} + S_{m+1,m+2} + \frac{m}{2(m+2)}\left( \frac{2}{m} T_{1,m} - T_{m+1,m+2} \right)^2.\]

where $T_{m+1,m+2}$ is the pairwise Tvalue for the PairAccumulator.

OnlineLogBinning.Svalue — Method

Svalue(pacc::PairAccumulator)

The $S$ function for a single pair following the accumulation of $m$ data points follows as

\[\begin{aligned} S_{m+1, m+2} &\equiv \sum_{k = m+1}^{m+2} \left( x_k - \frac{1}{2} T_{m+1,m+2} \right)^2 \\ &= \frac{1}{2}\left( x_{m+2} - x_{m+1} \right)^2. \end{aligned}\]

Thus, $S_{m+1,m+2}$ does not need to take $T_{m+1,m+2}$ as an argument.

OnlineLogBinning.Tvalue — Method

Tvalue(lacc::LevelAccumulator)

Function to calculate the online $T_{1,m+2}$ summation as:

\[T_{1,m+2} = T_{1,m} + T_{m+1,m+2},\]

where $T_{m+1,m+2}$ is the pairwise Tvalue for the PairAccumulator.

OnlineLogBinning.Tvalue — Method

Tvalue(pacc::PairAccumulator)

The $T$ function for a single pair following the accumulation of $m$ data points follows as

\[T_{m+1, m+2} \equiv \sum_{k = m+1}^{m+2} x_k = x_{m+1} + x_{m+2},\]

as expected.

OnlineLogBinning.bin_depth — Method

bin_depth(bacc::BinningAccumulator)

Number of binned levels present. length of the [BinningAccumulator] minus 1.

OnlineLogBinning.binning_level — Method

binning_level(index::Int)

Conversion from LvlAccums index to binning_level.

OnlineLogBinning.reset! — Method

reset!(bacc::BinningAccumulator{T})

Reset the BinningAccumulator by reconstruction.

Additional information

While this is not a literal reset per se, with a large enough BinningAccumulator it will be certainly faster just to blow up the old one (in memory) and start over.

OnlineLogBinning.reset! — Method

reset!(pacc::PairAccumulator)

Return the PairAccumulator to its initial state. Presumably one just exported the Tvalue and Svalue from it before the reset!.

OnlineLogBinning.std_error — Method

std_error( bacc::BinningAccumulator )

Online measurement of the [BinningAccumulator] standard error.

Additional information

This quantity is considered online despite that it is not regularly updated when data is push!ed from the stream.

OnlineLogBinning.std_error — Method

std_error( lacc::LevelAccumulator ) = sqrt(var_of_mean(lacc))

Online measurement of the [LevelAccumulator] standard error.

Additional information

This quantity is considered online despite that it is not regularly updated when data is push!ed from the stream.

OnlineLogBinning.update_SandT! — Method

update_SandT!(lacc::LevelAccumulator)

Apply the Svalue and Tvalue formula to update lacc.Taccum and return the PairAccumulator Tvalue increment.

Additional information

$S$ must be updated before $T$ since the former depends on the latter's history.

OnlineLogBinning.update_num_bins! — Function

update_num_bins!(lacc::LevelAccumulator, [incr = 2])

Increment the number of bins accumulated by incr.

OnlineLogBinning.var_of_mean — Method

var_of_mean( bacc::BinningAccumulator; [level = 0] )

Online measurement of the [BinningAccumulator] variance of the mean.

Additional information

This quantity is considered online despite that it is not regularly updated when data is push!ed from the stream.

OnlineLogBinning.var_of_mean — Method

var_of_mean( lacc::LevelAccumulator ) = var(lacc) / lacc.num_bins

Online measurement of the [LevelAccumulator] variance of the mean.

Additional information

This quantity is considered online despite that it is not regularly updated when data is push!ed from the stream.

Statistics.mean — Method

mean( bacc::BinningAccumulator; [level = 0] )

Online measurement of the [BinningAccumulator] mean.

Additional information

This quantity is considered online despite that it is not regularly updated when data is push!ed from the stream.

Statistics.mean — Method

mean( lacc::LevelAccumulator )

Online measurement of the [LevelAccumulator] mean.

Additional information

This quantity is considered online despite that it is not regularly updated when data is push!ed from the stream.

Statistics.std — Method

std( bacc::BinningAccumulator )

Online measurement of the [BinningAccumulator] standard deviation.

Additional information

This quantity is considered online despite that it is not regularly updated when data is push!ed from the stream.

Statistics.std — Method

std( lacc::LevelAccumulator ) = sqrt(var(lacc))

Online measurement of the [LevelAccumulator] standard deviation.

Additional information

This quantity is considered online despite that it is not regularly updated when data is push!ed from the stream.

Statistics.var — Method

var( bacc::LevelAccumulator; [level = 0] )

Online measurement of the [BinningAccumulator] variance.

Additional information

This quantity is considered online despite that it is not regularly updated when data is push!ed from the stream.

Statistics.var — Method

var( lacc::LevelAccumulator )

Online measurement of the [LevelAccumulator] variance.

Additional information

This quantity is considered online despite that it is not regularly updated when data is push!ed from the stream.